Introduce PlainOldDBO class as abstraction of all DBOs
