+++ app/src/main/java/justbe/mindfulnessapp/models/PlainOldDBO.java
@@ -1,0 +1,36 @@
+package justbe.mindfulnessapp.models;
+
+/**
+ * Created by eddiehurtig on 11/30/15.
+ */
+public abstract class PlainOldDBO<T> extends BaseModel<T> {
+    private Integer id;
+    private String created_at;
+    private String updated_at;
+
+    public String getCreated_at() {
+        return created_at;
+    }
+
+    public void setCreated_at(String created_at) {
+        this.created_at = created_at;
+    }
+
+    public String getUpdated_at() {
+        return updated_at;
+    }
+
+    public void setUpdated_at(String updated_at) {
+        this.updated_at = updated_at;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+}
+
